64bit Adaption perlscript

+++ opensource/kernel-cpp/zschimmer/perl.cxx
@@ -167,7 +167,7 @@ void Perl::parse( const string& script_text, VARIANT* result )
     
         args[2] = script_text.c_str();
 
-        fprintf( stderr, "perl_parse _perl=%X  %s,%s,%s\n", (int)PERL, args[0], args[1], args[2] );
+        fprintf( stderr, "perl_parse _perl=%X  %s,%s,%s\n", (int64)PERL, args[0], args[1], args[2] );
 Z_LOG2( "", "perl_parse(): " << args[2] << "\n" );
         err = perl_parse( PERL, xs_init, 3, (char**)args, NULL );    if(err)  throw_perl( err, "perl_parse" );
 Z_LOG2( "", "perl_parse() OK\n" );
@@ -214,7 +214,7 @@ void Perl::start( bool start_ )
         const char* args[] = { "", "-e", "" };
         args[2] = _script.c_str();
 
-      //fprintf( stderr, "perl_parse _perl=%X  %s,%s,%s\n", (int)PERL, args[0], args[1], args[2] );
+      //fprintf( stderr, "perl_parse _perl=%X  %s,%s,%s\n", (int64)PERL, args[0], args[1], args[2] );
       //err = perl_parse( PERL, xs_init, 3, (char**)args, NULL );    if(err)  throw_perl(err,"perl_parse"); //string_from_pv(ERRSV).c_str());
 Z_LOG2( "", "perl_parse(): " << _script.c_str() << "\n" );
         err = perl_parse( PERL, xs_init, 3, (char**)args, NULL );
